Rename `controllers` to `controller`
