+++ internal/controller/source_predicate.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers
+package controller
 
 import (
 	"sigs.k8s.io/controller-runtime/pkg/event"
